+++ source/controllers/src/impedance/CartesianTwistController.cpp
@@ -16,13 +16,12 @@ CartesianTwistController::CartesianTwistController(double linear_principle_dampi
     angular_damping_(std::make_shared<Parameter<double>>("angular_damping", angular_damping)),
     dissipative_ctrl_(ComputationalSpaceType::LINEAR),
     velocity_impedance_ctrl_(Eigen::MatrixXd::Zero(6, 6), Eigen::MatrixXd::Zero(6, 6)) {
+  set_linear_gains(linear_principle_damping, linear_orthogonal_damping);
   set_angular_gains(angular_stiffness, angular_damping);
 }
 
 CartesianTwistController::CartesianTwistController(const Eigen::Vector4d& gains) :
-    CartesianTwistController(gains(0), gains(1), gains(2), gains(3)) {
-
-}
+    CartesianTwistController(gains(0), gains(1), gains(2), gains(3)) {}
 
 CartesianTwistController::CartesianTwistController(const CartesianTwistController& other) :
     linear_principle_damping_(std::make_shared<Parameter<double>>("linear_principle_damping", other.get_gains()(0))),
@@ -30,8 +29,7 @@ CartesianTwistController::CartesianTwistController(const CartesianTwistControlle
     angular_stiffness_(std::make_shared<Parameter<double>>("angular_stiffness", other.get_gains()(2))),
     angular_damping_(std::make_shared<Parameter<double>>("angular_damping", other.get_gains()(3))),
     dissipative_ctrl_(other.dissipative_ctrl_),
-    velocity_impedance_ctrl_(other.velocity_impedance_ctrl_) {
-}
+    velocity_impedance_ctrl_(other.velocity_impedance_ctrl_) {}
 
 CartesianTwistController& CartesianTwistController::operator=(const CartesianTwistController& other) {
   CartesianTwistController tmp(other);
@@ -91,7 +89,8 @@ void CartesianTwistController::set_angular_gains(double angular_stiffness, doubl
   Eigen::MatrixXd k(6, 6), d(6, 6);
   k.diagonal() << 0, 0, 0, angular_stiffness, angular_stiffness, angular_stiffness;
   d.diagonal() << 0, 0, 0, angular_damping, angular_damping, angular_damping;
-  velocity_impedance_ctrl_ = VelocityImpedance<CartesianState>(k, d);
+  velocity_impedance_ctrl_.set_stiffness(k);
+  velocity_impedance_ctrl_.set_damping(d);
 }
 
 std::list<std::shared_ptr<ParameterInterface>> CartesianTwistController::get_parameters() const {